Renamed file, removed extra variables
